backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.0
python-dotenv==1.0.1
